refactor to abstract Box and 2 subclassses
We made an abstract Box class which contains the 4 sides of a box. This
is subclassed into BoundingBox and BoundingBoxRange, which different
behavior. If you expand a BoundingBox, e.g., this adds and subtracts
some values.
If you expand a BoundinBoxRange however, this should only add the
values.

This error was made and to make sure it doesn't happen again, we split
them into two classes which differ in behavior.